+++ src/OpenTelemetry.ResourceDetectors.AWS/AWSResourcesEventSource.cs
@@ -24,12 +24,12 @@ internal sealed class AWSResourcesEventSource : EventSource
     [Event(1, Message = "Failed to extract resource attributes in '{0}'.", Level = EventLevel.Warning)]
     public void FailedToExtractResourceAttributes(string format, string exception)
     {
-        this.WriteEvent(3, format, exception);
+        this.WriteEvent(1, format, exception);
     }
 
     [Event(2, Message = "Failed to validate certificate in format: '{0}', error: '{1}'.", Level = EventLevel.Warning)]
     public void FailedToValidateCertificate(string format, string error)
     {
-        this.WriteEvent(4, format, error);
+        this.WriteEvent(2, format, error);
     }
 }